Add a way to configure Grafana 2
In a discussion with @bfraser, we decided to introduce a single
configuration hash parameter. This allows for any configuration option
to be set without relasing a new version (which would be necessary when
creating a parameter for each settings). So this allows for more
flexicility.

Also we decided on using no defaults, since Grafana has sane defaults
already. So adding our own defaults would only complicate things.