add parameter install_with_apt
which defines if the package is installed with the apt provider. This gives
you the possibility to install the grafana package from the apt sources.
This is especially useful if you already manage apt repos with puppet, e.g.
with the puppetlabs-apt module.

The default value is set to false so it doesn't change the existent behaviour
of this module.