Add acceptance testing for teams, bugfix dashboard
When searching for a dashboard, it would not check the current
organization. This could cause dashboards to not be found if on
a different organization.

Have added centos 7 hostfile with some info on how to run beaker tests.
Acceptance testing to see if teams can be created, updated and torn down.