Drop older grafana versions (6.x, 7.x) from tests
The last 7.x release was ~6 month ago and basically correspond to the
beginning of availability of version 9.x.  7.x not being updated
anymore, we can drop it (and older versions) with the benefit of
speeding-up the acceptance test suite.

While here, bump the version of the older releases which are older than
the operating systems we support and where sometimes not compatible.  Do
not do this for 9.x because we also test with 'latest'.